Fixes to course & section creation

+++ client/app/controllers/course-version.js
@@ -1,27 +1,29 @@
-import Ember from 'ember';
+import Ember from "ember";
 
 export default Ember.ObjectController.extend({
   isUnderReview: false,
   actions: {
     saveModel: function () {
-      this.get('model').save();
+      this.get("model").save();
     },
     deleteModel: function () {
-      this.get('model').deleteRecord();
-      this.get('model').save();
-      this.transitionToRoute('courseVersions');
+      if (confirm("Are you sure you wish to delete this course version?")) {
+        this.get("model").deleteRecord();
+        this.get("model").save();
+        this.transitionToRoute("courseVersions");
+      }
     },
     startReview: function () {
-      this.set('isUnderReview', true);
+      this.set("isUnderReview", true);
     },
     finalizeReview: function () {
-      this.set('isUnderReview', false);
-      var thisModel = this.get('model');
-      var newReview = this.store.createRecord('qualityReview', {
+      this.set("isUnderReview", false);
+      var thisModel = this.get("model");
+      var newReview = this.store.createRecord("qualityReview", {
         courseVersion: thisModel,
         startDate: new Date(),
         endDate: new Date(),
-        stage: 'Complete'
+        stage: "Complete"
       });
       newReview.save();
     }
